requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Django==4.0.2
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
password==0.2
pytz==2021.3
sqlparse==0.4.2
